Fix for crash with an FPU: converting a float to a int32, then storing it as a int16.

+++ g2core/config.h
@@ -170,7 +170,8 @@
 
 // Sizing and footprints                // chose one based on # of elements in cfgArray
 //typedef uint8_t index_t;              // use this if there are < 256 indexed objects
-typedef uint16_t index_t;               // use this if there are > 255 indexed objects
+//typedef uint16_t index_t;               // use this if there are > 255 indexed objects
+typedef uint32_t index_t;               // use this because set/get_int is expecting an int32_t
 
                                         // defines allocated from stack (not-pre-allocated)
 #define NV_FORMAT_LEN 128               // print formatting string max length
